remove root empiric
